feat: create and throw exceptions

+++ backend/gastroit/src/main/java/com/example/gastroit/service/UnitServiceImpl.java
@@ -5,6 +5,8 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.example.gastroit.exception.UnitAlreadyExistException;
+import com.example.gastroit.exception.UnitNotFoundException;
 import com.example.gastroit.model.entity.Unit;
 import com.example.gastroit.repository.UnitRepository;
 
@@ -16,11 +18,17 @@ public class UnitServiceImpl implements UnitService {
 
 	@Override
 	public Unit create(Unit unit) {
+		if (list().stream().anyMatch(u -> u.getName().equals(unit.getName()))) {
+			throw new UnitAlreadyExistException();
+		}
 		return unitRepository.save(unit);
 	}
 
 	@Override
 	public void delete(Long id) {
+		if (list().stream().noneMatch(u -> u.getId().equals(id))) {
+			throw new UnitNotFoundException();
+		}
 		unitRepository.deleteById(id);
 	}
